- Update on Bin and libs:     * Ogre 1.7.1     * CEGUI 0.7.1     * Bullet 2.76     * OgreBullet (last svn) - Updates on includes for Ogre and CEGUI in property sheets - Coordinate system NORMNAL -> NORMAL3D - CameraController now uses Ogre Camera Control System library - Improvements in Physics (disabled by default) - Fixes in project properties: now all use property sheets - Small changes in physics demo

+++ cing/src/common/MathUtils.h
@@ -28,7 +28,7 @@
 #include "PerlinNoise.h"
 #include "framework/UserAppGlobals.h"
 
-#include "Ogre3d/include/OgreBitwise.h"
+#include "OgreBitwise.h"
 
 #include <numeric>
 # include <cmath>
@@ -383,7 +383,7 @@ struct Average
 	}
 
 	/// @brief returns the ave
-	float getValue()
+	double getValue()
 	{
 		double sum = std::accumulate( values.begin(), values.end(), 0.0 );
 		return sum / (double)values.size();
